requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-api-key==2.3.0
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
